Added ElasticSearch for games search

+++ src/HorCup.Infrastructure/Responses/PagedSearchResponse.cs
@@ -4,5 +4,5 @@ using System.Diagnostics.CodeAnalysis;
 namespace HorCup.Infrastructure.Responses
 {
 	[ExcludeFromCodeCoverage]
-	public record PagedSearchResponse<T>(IEnumerable<T> Items, int Total);
+	public record PagedSearchResponse<T>(IEnumerable<T> Items, long Total);
 }
